+++ packages/asc-ui/src/components/Radio/RadioItemStyle.ts
@@ -1,6 +1,0 @@
-import styled from '@datapunt/asc-core'
-
-const RadioItemStyle = styled.div`
-  position: relative;
-`
-export default RadioItemStyle