+++ packages/asc-ui/src/components/Radio/index.ts
@@ -1,3 +1,3 @@
-import * as Radio from './Radio'
-
-export default Radio
+export { default as RadioStyle, RadioWrapperStyle } from './RadioStyle'
+export { default } from './Radio'
+export { default as RadioGroup } from './RadioGroup'